Divide receive::v1 into exclusive and shared mods
The exclusive module is used only when v1 is explicitly enabled. Shared components
are `pub(crate)` and used when the v2 feature is enabled but not exported.
