+++ payjoin/Cargo.toml
@@ -18,15 +18,17 @@ exclude = ["tests"]
 [features]
 default = ["v2"]
 base64 = ["bitcoin/base64"]
-v1 = ["bitcoin/rand"]
-v2 = ["bitcoin/rand", "bitcoin/serde", "hpke", "dep:http", "bhttp", "ohttp", "serde", "url/serde" ]
+#[doc = "Core features for payjoin state machines"]
+_core = ["bitcoin/rand", "serde_json", "url", "bitcoin_uri"]
+v1 = ["_core"]
+v2 = ["_core", "bitcoin/serde", "hpke", "dep:http", "bhttp", "ohttp", "serde", "url/serde"]
 #[doc = "Functions to fetch OHTTP keys via CONNECT proxy using reqwest. Enables `v2` since only `v2` uses OHTTP."]
 io = ["v2", "reqwest/rustls-tls"]
 _danger-local-https = ["reqwest/rustls-tls", "rustls"]
 
 [dependencies]
 bitcoin = { version = "0.32.5", features = ["base64"] }
-bitcoin_uri = "0.1.0"
+bitcoin_uri = { version = "0.1.0", optional = true }
 hpke = { package = "bitcoin-hpke", version = "0.13.0", optional = true }
 log = { version = "0.4.14"}
 http = { version = "1", optional = true }
@@ -35,8 +37,8 @@ ohttp = { package = "bitcoin-ohttp", version = "0.6.0", optional = true }
 serde = { version = "1.0.186", default-features = false, optional = true }
 reqwest = { version = "0.12", default-features = false, optional = true }
 rustls = { version = "0.22.4", optional = true }
-url = "2.2.2"
-serde_json = "1.0.108"
+url = { version = "2.2.2", optional = true }
+serde_json = { version = "1.0.108", optional = true }
 
 [dev-dependencies]
 bitcoind = { version = "0.36.0", features = ["0_21_2"] }
