+++ payjoin/src/receive/mod.rs
@@ -8,7 +8,11 @@ use crate::psbt::{InternalInputPair, InternalPsbtInputError};
 mod error;
 pub(crate) mod optional_parameters;
 
+#[cfg(feature = "v1")]
 pub mod v1;
+#[cfg(not(feature = "v1"))]
+pub(crate) mod v1;
+
 #[cfg(feature = "v2")]
 pub mod v2;
